jit, gc: drop the dead compiled-loop unwrap, widen the drain decile scale
Two findings from the PR review bots on 945.

`run_compiled_raw_with_values` bound
`self.compiled_loops.get(&green_key).unwrap()` right after
`execute_token_ints` and never read it — the later `.and_then(|compiled|
…)` shadows the name inside its own closure.  The binding is dead but
the `unwrap()` is not: a residual call that re-enters the driver can
remove this green key during the run, and the panic then fires ahead of
the fresh fallible lookup and its `trace_for_exit_by_rd_loop_token`
fallback, so that fallback could never run in the case it was written
for.  Deleting the binding leaves no other use.

`record_drain_sample` scaled the decile with `promoted.min(used_before)
* 100` in `usize`.  On wasm32 that wraps once the nursery passes ~43 MB,
bucketing the sample wrong in release and tripping the overflow check in
debug.  Widen to u64 before scaling.

pyre/check.py --synthetic-only --backend dynasm: 356/356.

+++ majit/majit-gc/src/lib.rs
@@ -177,7 +177,10 @@ pub fn drain_census_record(
     let percent = if used_before == 0 {
         0
     } else {
-        (promoted.min(used_before) * 100) / used_before
+        // Widen before scaling: on wasm32 a nursery past ~43 MB makes
+        // `used_before * 100` wrap a 32-bit `usize`, which buckets the sample
+        // wrong in release and trips the overflow check in debug.
+        ((promoted.min(used_before) as u64 * 100) / used_before as u64) as usize
     };
     census.deciles[(percent / 10).min(DRAIN_DECILES - 1)] += 1;
     if drain_census_dump_interval()
